Clean up legacy pages routes
