--- assets/todo.css
@@ -1,349 +1,0 @@
-/* Todo Kanban Board Styles */
-/* Uses the same CSS variables as style.css for theme consistency */
-
-.todo-board {
-  max-width: 1100px;
-  margin: 0 auto;
-  padding: 24px;
-}
-
-.todo-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 24px;
-}
-
-.todo-title {
-  font-size: 1.75em;
-  font-weight: 600;
-  color: var(--color-text);
-}
-
-.todo-add-btn {
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-  padding: 8px 16px;
-  font-size: 14px;
-  font-weight: 600;
-  border: none;
-  border-radius: 6px;
-  background: var(--color-link);
-  color: #fff;
-  cursor: pointer;
-}
-
-.todo-add-btn:hover { opacity: 0.9; }
-
-/* Columns */
-.todo-columns {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 24px;
-  align-items: start;
-}
-
-.todo-column {
-  min-height: 120px;
-}
-
-.todo-column-title {
-  font-size: 1.1em;
-  font-weight: 600;
-  color: var(--color-text);
-  margin-bottom: 12px;
-  padding-bottom: 8px;
-  border-bottom: 2px solid var(--color-border);
-}
-
-.todo-count {
-  font-size: 0.8em;
-  font-weight: 400;
-  color: var(--color-text-secondary);
-  background: var(--color-code-bg);
-  padding: 2px 8px;
-  border-radius: 10px;
-  margin-left: 4px;
-}
-
-.todo-cards {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.todo-empty {
-  font-size: 14px;
-  color: var(--color-text-secondary);
-  font-style: italic;
-  padding: 16px;
-  text-align: center;
-  border: 1px dashed var(--color-border);
-  border-radius: 8px;
-}
-
-/* Cards */
-.todo-card {
-  border: 1px solid var(--color-border);
-  border-radius: 8px;
-  padding: 14px 16px;
-  background: var(--color-bg);
-  transition: box-shadow 0.15s ease;
-}
-
-.todo-card:hover {
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-}
-
-[data-theme="dark"] .todo-card:hover {
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.3);
-}
-
-.todo-card-header {
-  display: flex;
-  align-items: baseline;
-  gap: 8px;
-  margin-bottom: 6px;
-}
-
-.todo-card-id {
-  font-size: 12px;
-  font-family: var(--font-mono);
-  color: var(--color-text-secondary);
-  flex-shrink: 0;
-}
-
-.todo-card-title {
-  font-size: 15px;
-  font-weight: 500;
-  color: var(--color-text);
-  line-height: 1.4;
-}
-
-/* Metadata */
-.todo-card-meta {
-  margin-top: 8px;
-  padding-top: 8px;
-  border-top: 1px solid var(--color-border);
-}
-
-.todo-meta-line {
-  font-size: 13px;
-  color: var(--color-text-secondary);
-  line-height: 1.5;
-  word-break: break-word;
-}
-
-.todo-meta-line strong {
-  color: var(--color-text-secondary);
-  font-weight: 600;
-}
-
-.todo-meta-line a {
-  color: var(--color-link);
-  text-decoration: none;
-}
-
-.todo-meta-line a:hover {
-  text-decoration: underline;
-}
-
-/* Action buttons */
-.todo-card-actions {
-  display: flex;
-  gap: 6px;
-  margin-top: 10px;
-  padding-top: 8px;
-  border-top: 1px solid var(--color-border);
-}
-
-.todo-action-btn {
-  padding: 4px 10px;
-  font-size: 13px;
-  border-radius: 4px;
-  cursor: pointer;
-  border: 1px solid var(--color-border);
-  background: none;
-  color: var(--color-text-secondary);
-  transition: all 0.15s ease;
-}
-
-.todo-action-btn:hover {
-  background: var(--color-code-bg);
-  color: var(--color-text);
-}
-
-.todo-complete-btn:hover {
-  border-color: #2da44e;
-  color: #2da44e;
-  background: rgba(45, 164, 78, 0.08);
-}
-
-.todo-reopen-btn:hover {
-  border-color: var(--color-link);
-  color: var(--color-link);
-  background: rgba(9, 105, 218, 0.08);
-}
-
-.todo-delete-btn:hover {
-  border-color: #d1242f;
-  color: #d1242f;
-  background: rgba(209, 36, 47, 0.08);
-}
-
-/* Completed column cards are slightly muted */
-.todo-column-completed .todo-card {
-  opacity: 0.75;
-}
-
-.todo-column-completed .todo-card:hover {
-  opacity: 1;
-}
-
-/* Add Item Modal */
-.todo-modal {
-  position: fixed;
-  inset: 0;
-  z-index: 1000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.todo-modal[hidden] { display: none; }
-
-.todo-modal-backdrop {
-  position: absolute;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.5);
-}
-
-.todo-modal-content {
-  position: relative;
-  background: var(--color-bg);
-  border: 1px solid var(--color-border);
-  border-radius: 12px;
-  width: 480px;
-  max-width: 90vw;
-  max-height: 80vh;
-  overflow-y: auto;
-  box-shadow: 0 8px 30px rgba(0, 0, 0, 0.12);
-}
-
-.todo-modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px 20px;
-  border-bottom: 1px solid var(--color-border);
-}
-
-.todo-modal-header h3 {
-  font-size: 16px;
-  font-weight: 600;
-  margin: 0;
-}
-
-.todo-modal-close {
-  background: none;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-  color: var(--color-text-secondary);
-  padding: 0 4px;
-  line-height: 1;
-}
-
-.todo-modal-close:hover { color: var(--color-text); }
-
-.todo-modal-body {
-  padding: 20px;
-}
-
-.todo-modal-body label {
-  display: block;
-  font-size: 13px;
-  color: var(--color-text-secondary);
-  margin-bottom: 4px;
-  margin-top: 12px;
-}
-
-.todo-modal-body label:first-child {
-  margin-top: 0;
-}
-
-.todo-input {
-  width: 100%;
-  padding: 8px 12px;
-  font-size: 14px;
-  border: 1px solid var(--color-border);
-  border-radius: 6px;
-  background: var(--color-bg);
-  color: var(--color-text);
-  box-sizing: border-box;
-}
-
-.todo-input:focus {
-  outline: none;
-  border-color: var(--color-link);
-}
-
-.todo-textarea {
-  width: 100%;
-  padding: 8px 12px;
-  font-size: 14px;
-  border: 1px solid var(--color-border);
-  border-radius: 6px;
-  background: var(--color-bg);
-  color: var(--color-text);
-  box-sizing: border-box;
-  resize: vertical;
-  font-family: var(--font-body);
-}
-
-.todo-textarea:focus {
-  outline: none;
-  border-color: var(--color-link);
-}
-
-.todo-submit-btn {
-  width: 100%;
-  margin-top: 16px;
-  padding: 10px;
-  font-size: 14px;
-  font-weight: 600;
-  border: none;
-  border-radius: 6px;
-  background: var(--color-link);
-  color: #fff;
-  cursor: pointer;
-}
-
-.todo-submit-btn:hover { opacity: 0.9; }
-.todo-submit-btn:disabled { opacity: 0.5; cursor: not-allowed; }
-
-/* Responsive */
-@media (max-width: 768px) {
-  .todo-board {
-    padding: 16px;
-  }
-
-  .todo-header {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 12px;
-  }
-
-  .todo-title {
-    font-size: 1.4em;
-  }
-
-  .todo-columns {
-    grid-template-columns: 1fr;
-  }
-
-  .todo-modal-content {
-    width: 95vw;
-  }
-}
